Fixed logged in messages and options

+++ core/models.py
@@ -5,7 +5,7 @@ from django.utils import timezone
 
 
 class Campus(models.Model):
-    id = models.IntegerField(primary_key=True)
+    id = models.PositiveSmallIntegerField(primary_key=True)
     nombre = models.CharField(max_length=100)
     imagen = models.ImageField(null=False, blank=False)
 
@@ -13,7 +13,7 @@ class Campus(models.Model):
 	    return self.nombre
 
 class Division(models.Model):
-    id = models.IntegerField(primary_key=True)
+    id = models.PositiveSmallIntegerField(primary_key=True)
     campus = models.ForeignKey(Campus, on_delete=models.CASCADE)
     nombre = models.CharField(max_length=100)
     imagen = models.ImageField(null=False, blank=False)
